+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog for aim.models
 - Added default to IS3Bucket.policy
 - Added Route53 to schema and model
 - Removed unused yaml config from aimdemo under fixtures.
+- Ported CodeCommit to schema and model
 
 
 1.0.1 (2019-06-19)